+++ packages/skyline-ts/src/providers/index.ts
@@ -1,14 +1,11 @@
-export {
-  dedicatedLines,
-  imessage,
-} from "@skyline-ts/imessage";
+export { imessage } from "@skyline-ts/imessage";
 export type {
   DedicatedLine,
   ImessageCloudConfig,
   ImessageConfig,
   ImessageDedicatedConfig,
 } from "@skyline-ts/imessage";
-export { slack, slackDedicatedLines } from "@skyline-ts/slack";
+export { slack } from "@skyline-ts/slack";
 export type {
   SlackCloudConfig,
   SlackConfig,
@@ -18,17 +15,14 @@ export type {
 } from "@skyline-ts/slack";
 export { terminal } from "@skyline-ts/terminal";
 export type { TerminalConfig } from "@skyline-ts/terminal";
-export { whatsapp, whatsappDedicatedLines } from "@skyline-ts/whatsapp";
+export { whatsapp } from "@skyline-ts/whatsapp";
 export type {
   WhatsappCloudConfig,
   WhatsappConfig,
   WhatsappDedicatedConfig,
   WhatsappDedicatedLine,
 } from "@skyline-ts/whatsapp";
-export {
-  whatsappBusiness,
-  whatsappBusinessDedicatedLines,
-} from "@skyline-ts/whatsapp-business";
+export { whatsappBusiness } from "@skyline-ts/whatsapp-business";
 export type {
   WhatsappBusinessCloudConfig,
   WhatsappBusinessConfig,