+++ packages/slack/src/index.ts
@@ -1,7 +1,6 @@
 export { bind } from "./bind.js";
 export {
   slack,
-  slackDedicatedLines,
   type SlackCloudConfig,
   type SlackConfig,
   type SlackDedicatedConfig,